+++ src/core/io/ExportFileWriter.ts
@@ -11,10 +11,20 @@ import { AeroflyFlightToKmlConverter } from "../converter/aerofly-flight/Aerofly
  * external flight plan file by selecting the appropriate converter.
  */
 export class ExportFileWriter {
-    static exportFlightplanToFile(filename: string, flightplan: AeroflyFlight): void {
-        const converter = this.getConverter(filename);
-        const content = new converter().convert(flightplan);
+    static fileTypes: string[] = [
+        AeroflyFlightToAeroflyMainMcfConverter.fileExtension,
+        AeroflyFlightToAeroflyCustomMissionsTmcConverter.fileExtension,
+        AeroflyFlightToGeoJsonConverter.fileExtension,
+        AeroflyFlightToKmlConverter.fileExtension,
+    ] as const;
 
+    static exportFlightplanToString(filename: string, flightplan: AeroflyFlight): string {
+        const converter = this.getConverter(filename);
+        return new converter().convert(flightplan);
+    }
+
+    static exportFlightplanToFile(filename: string, flightplan: AeroflyFlight): void {
+        const content = ExportFileWriter.exportFlightplanToString(filename, flightplan);
         fs.writeFileSync(filename, content, "utf8");
     }