+++ src/core/io/ImportFileReader.ts
@@ -6,28 +6,33 @@ import { StringToAeroflyFlightConverter } from "../converter/other/StringToAerof
 import { XplaneFmsToAeroflyFlightConverter } from "../converter/other/XplaneFmsToAeroflyFlightConverter.js";
 import { AeroflyMcfToImportFileConverter } from "../converter/other/AeroflyMcfToImportFileConverter.js";
 import { AeroflyCustomMissionsTmcToAeroflyFlightConverter } from "../converter/other/AeroflyCustomMissionsTmcToAeroflyFlightConverter.js";
-import { AeroflyCustomMissionsJsonToAeroflyFlightConverter } from "../converter/other/AeroflyCustomMissionsJsonToAeroflyFlightConverter.js";
 
 /**
  * Reads a file and converts it into `AeroflyFlight` by selecting the
  * appropriate converter class.
  */
 export class ImportFileReader {
+    static fileTypes: string[] = [
+        AeroflyCustomMissionsTmcToAeroflyFlightConverter.fileExtension,
+        AeroflyMcfToImportFileConverter.fileExtension,
+        MsfsPlnToAeroflyFlightConverter.fileExtension,
+        GarminFplToAeroflyFlightConverter.fileExtension,
+        XplaneFmsToAeroflyFlightConverter.fileExtension,
+    ] as const;
+
     static getFlightplansFromFile(filename: string): string[] {
         const content = fs.readFileSync(filename, "utf8");
         return this.getFlightplansFromString(content, filename);
     }
 
-    static getFlightplansFromString(content: string, filename: string) {
+    static getFlightplansFromString(content: string, filename: string): string[] {
         const converter = this.getConverter(filename);
         return new converter().getIndices(content);
     }
 
     /**
      * Imports a flight plan from a file and converts it to an AeroflyFlight object.
-     * Supported file types are determined by the file extension:
-     * - .pln: Microsoft Flight Simulator flight plan file
-     * - .fpl: Garmin FPL file
+     * Supported file types are determined by the file extension.
      *
      * @param filename The path to the flight plan file to import.
      * @param flightplan The AeroflyFlight object to populate with the imported data.
@@ -67,8 +72,6 @@ export class ImportFileReader {
 
     static getRegistry(): Record<string, (new () => StringToAeroflyFlightConverter) | undefined> {
         return {
-            [AeroflyCustomMissionsJsonToAeroflyFlightConverter.fileExtension]:
-                AeroflyCustomMissionsJsonToAeroflyFlightConverter,
             [AeroflyCustomMissionsTmcToAeroflyFlightConverter.fileExtension]:
                 AeroflyCustomMissionsTmcToAeroflyFlightConverter,
             [AeroflyMcfToImportFileConverter.fileExtension]: AeroflyMcfToImportFileConverter,