+++ src/core/services/ImportFileFinderService.ts
@@ -6,7 +6,6 @@ import { MsfsPlnToAeroflyFlightConverter } from "../converter/other/MsfsPlnToAer
 import { XplaneFmsToAeroflyFlightConverter } from "../converter/other/XplaneFmsToAeroflyFlightConverter.js";
 import { AeroflyMcfToImportFileConverter } from "../converter/other/AeroflyMcfToImportFileConverter.js";
 import { AeroflyCustomMissionsTmcToAeroflyFlightConverter } from "../converter/other/AeroflyCustomMissionsTmcToAeroflyFlightConverter.js";
-import { AeroflyCustomMissionsJsonToAeroflyFlightConverter } from "../converter/other/AeroflyCustomMissionsJsonToAeroflyFlightConverter.js";
 
 /**
  * Finds local flight plan files
@@ -27,7 +26,6 @@ export class ImportFileFinderService {
         const files = fs.readdirSync(importDirectory);
         const importFiles = files.filter(
             (file: string) =>
-                file.toLowerCase().endsWith(AeroflyCustomMissionsJsonToAeroflyFlightConverter.fileExtension) ||
                 file.toLowerCase().endsWith(AeroflyCustomMissionsTmcToAeroflyFlightConverter.fileExtension) ||
                 file.toLowerCase().endsWith(AeroflyMcfToImportFileConverter.fileExtension) ||
                 file.toLowerCase().endsWith(GarminFplToAeroflyFlightConverter.fileExtension) ||